+++ front/src/Phaser/Game/GameScene.ts
@@ -989,8 +989,8 @@ ${escapedMessage}
             })
         );
 
-        this.iframeSubscriptionList.push(
-            iframeListener.enablePlayerControlStream.subscribe(() => {
+       this.iframeSubscriptionList.push(
+            iframeListener.enablePlayerControlStream.subscribe(()=>{
                 this.userInputManager.restoreControls();
             })
         );
@@ -1058,7 +1058,13 @@ ${escapedMessage}
                     tags: this.connection ? this.connection.getAllTags() : [],
                 });
             })
-        );
+        )
+        this.iframeSubscriptionList.push(iframeListener.setTilesStream.subscribe((eventTiles) => {
+            for (const eventTile of eventTiles) {
+                this.gameMap.putTile(eventTile.tile, eventTile.x, eventTile.y, eventTile.layer);
+            }
+        }))
+
     }
 
     private setPropertyLayer(